refactor: delete dead compat wrappers, aliases, and duplicate exported names
Problem: the repo's stated policy is hard renames with no compat shims, yet
a sweep found surviving renamed-name wrappers and aliases — each one a
second spelling of a concept that queries/greps against the canonical name
silently miss. Two were live split-vocabulary hazards: VERB_NAMES was
exported under three names (VERB_NAMES / QUERY_VERB_NAMES / QUERY_VERBS)
and parser_diagnostics tested membership against two of them as if they
were different sets; cli/verb_names.QUERY_VERBS collided with an unrelated,
also-unreferenced cli/query_verbs.QUERY_VERBS tuple.

Deleted (all verified zero production callers by two grep spellings each):
- cli/query.py execute_query + async_execute_query param-shims (test-only;
  tests now drive async_execute_query_request directly)
- rendering build_rendered_message_payload wrapper + its core.py re-export
- rendering card_kind_for_tool wrapper (tests use classify_tool().card_kind)
- cli JudgeCandidateRow.from_claim compat constructor
- cli verb-name aliases QUERY_VERB_NAMES/QUERY_VERBS (callers on VERB_NAMES)
  and the dead query_verbs.QUERY_VERBS tuple
- cli/query_output dead aliases _output_summary_list/_open_result
- mcp/payloads stillborn MCP*-prefixed aliases (4, zero adopters)
- operations OperationStatus back-compat re-export (importers repointed to
  operations.operation_status, its real home)

Verification: devtools test on the five affected test files (302 passed);
devtools verify --quick exit 0.
