+++ polylogue/operations/operation_contract.py
@@ -13,14 +13,6 @@ already be imminent).
 
 What remains here is genuinely reusable, not speculative:
 
-* :class:`OperationStatus` — the scheduling status enum, now defined in
-  :mod:`polylogue.operations.operation_status` (polylogue-8s70) and re-exported
-  here for backward compatibility. Every member has a real consumer
-  (:mod:`polylogue.readiness.capability` maps ``RUNNING``/``COMPLETED``/
-  ``FAILED`` to capability readiness states) -- that consumer is on the hot
-  path of every ``polylogue status`` invocation and imports the enum from its
-  new zero-dependency home directly, to avoid paying for
-  :class:`OperationFollowUp`'s pydantic base.
 * :class:`OperationFollowUp` — the polling-hint shape returned by every
   accepted/pending operation. A plain data shape, not a polymorphic base —
   keeping it costs nothing and any second operation would want the same
@@ -34,7 +26,6 @@ abstract layer above it was removed.
 from __future__ import annotations
 
 from polylogue.core.json import JSONDocument, json_document
-from polylogue.operations.operation_status import OperationStatus
 from polylogue.surfaces.payloads import SurfacePayloadModel
 
 
@@ -71,5 +62,4 @@ class OperationFollowUp(SurfacePayloadModel):
 
 __all__ = [
     "OperationFollowUp",
-    "OperationStatus",
 ]
