+++ polylogue/rendering/core.py
@@ -7,14 +7,12 @@ from polylogue.rendering.core_markdown import format_session_markdown
 from polylogue.rendering.core_messages import (
     RenderedMessage,
     build_rendered_message,
-    build_rendered_message_payload,
 )
 
 __all__ = [
     "RenderableBlock",
     "RenderedMessage",
     "build_rendered_message",
-    "build_rendered_message_payload",
     "coerce_renderable_blocks",
     "format_session_markdown",
 ]
